+++ podcasts/ButtonCell.swift
@@ -2,5 +2,9 @@
 import UIKit
 
 class ButtonCell: ThemeableCell {
-    @IBOutlet var buttonTitle: UILabel!
+    @IBOutlet var buttonTitle: UILabel! {
+        didSet {
+            buttonTitle.font = UIFont.font(ofSize: 17.0, scalingWith: .body)
+        }
+    }
 }