+++ podcasts/PodcastImageView.swift
@@ -19,7 +19,8 @@ class PodcastImageView: UIView {
 
     func setPodcast(uuid: String, size: PodcastThumbnailSize) {
         guard let imageView = imageView else { return }
-
+        imageView.removeConstraints(imageView.constraints)
+        imageView.anchorToAllSidesOf(view: self)
         ImageManager.sharedManager.loadImage(podcastUuid: uuid, imageView: imageView, size: size, showPlaceHolder: true)
         adjustForSize(size)
     }