+++ podcasts/ProfileViewController.swift
@@ -11,15 +11,62 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
 
     @IBOutlet var footerView: UIView!
     @IBOutlet var alertIcon: UIImageView!
-    @IBOutlet var lastRefreshTime: UILabel!
-    @IBOutlet var refreshBtn: AnimatedImageButton! {
+    @IBOutlet var lastRefreshTime: ThemeableLabel! {
         didSet {
-            refreshBtn.mainColor = ThemeColor.primaryText02()
-            refreshBtn.buttonImage = UIImageView(image: UIImage(named: "profile-retry"))
+            lastRefreshTime.style = .primaryText02
+            lastRefreshTime.font = UIFont.font(with: .subheadline, maxSizeCategory: .accessibilityMedium)
+            lastRefreshTime.adjustsFontForContentSizeCategory = true
+        }
+    }
+    @IBOutlet var refreshButtonContainer: UIView!
 
-            refreshBtn.buttonTapped = { [weak self] in
+    private var refreshButtonTitle: String = L10n.refreshNow {
+        didSet {
+            updateRefreshButton()
+        }
+    }
+
+    private var isRefreshAnimating: Bool = false {
+        didSet {
+            updateRefreshButton()
+        }
+    }
+
+    private var refreshButtonHostingController: UIHostingController<AnyView>?
+
+    private func setupRefreshButton() {
+        updateRefreshButton()
+        if let hostingController = refreshButtonHostingController {
+            hostingController.sizingOptions = .intrinsicContentSize
+            addChild(hostingController)
+            refreshButtonContainer.addSubview(hostingController.view)
+            hostingController.view.translatesAutoresizingMaskIntoConstraints = false
+            NSLayoutConstraint.activate([
+                hostingController.view.topAnchor.constraint(equalTo: refreshButtonContainer.topAnchor),
+                hostingController.view.bottomAnchor.constraint(equalTo: refreshButtonContainer.bottomAnchor),
+                hostingController.view.centerXAnchor.constraint(equalTo: refreshButtonContainer.centerXAnchor),
+                hostingController.view.leadingAnchor.constraint(greaterThanOrEqualTo: refreshButtonContainer.leadingAnchor),
+                hostingController.view.trailingAnchor.constraint(lessThanOrEqualTo: refreshButtonContainer.trailingAnchor)
+            ])
+            hostingController.didMove(toParent: self)
+        }
+    }
+
+    private func updateRefreshButton() {
+        let refreshButton = ProfileRefreshButton(
+            title: refreshButtonTitle,
+            isAnimating: isRefreshAnimating,
+            action: { [weak self] in
                 self?.refreshTapped()
             }
+        ).setupDefaultEnvironment()
+
+        if let hostingController = refreshButtonHostingController {
+            hostingController.rootView = AnyView(refreshButton)
+        } else {
+            let hostingController = UIHostingController(rootView: AnyView(refreshButton))
+            hostingController.view.backgroundColor = .clear
+            self.refreshButtonHostingController = hostingController
         }
     }
 
@@ -90,6 +137,7 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
 
         profileTable.tableFooterView = footerView
 
+        setupRefreshButton()
         updateDisplayedData()
         updateRefreshFooterColors()
         updateFooterFrame()
@@ -165,8 +213,6 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
     }
 
     private func updateRefreshFooterColors() {
-        refreshBtn.mainColor = ThemeColor.primaryText02()
-        lastRefreshTime.textColor = ThemeColor.primaryText02()
         alertIcon.tintColor = ThemeColor.primaryIcon02()
     }
 
@@ -193,7 +239,7 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
     private func refreshTapped() {
         Analytics.track(.profileRefreshButtonTapped)
 
-        refreshBtn.animateImage(animationType: .rotate)
+        isRefreshAnimating = true
         lastRefreshTime.text = L10n.refreshing
         RefreshManager.shared.refreshPodcasts()
     }
@@ -205,7 +251,7 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
             guard let self = self else { return }
 
             self.refreshControl?.endRefreshing(true)
-            self.refreshBtn.stopAnimatingImage()
+            self.isRefreshAnimating = false
             self.updateLastRefreshDetails()
         }
     }
@@ -237,10 +283,10 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
 
         if !ServerSettings.lastRefreshSucceeded() || !ServerSettings.lastSyncSucceeded() {
             lastRefreshTime.text = !ServerSettings.lastRefreshSucceeded() ? L10n.refreshFailed : L10n.syncFailed
-            refreshBtn.buttonTitle = L10n.tryAgain
+            refreshButtonTitle = L10n.tryAgain
             alertIcon.isHidden = false
         } else if let lastUpdateTime = ServerSettings.lastRefreshEndTime() {
-            refreshBtn.buttonTitle = L10n.refreshNow
+            refreshButtonTitle = L10n.refreshNow
             if abs(lastUpdateTime.timeIntervalSinceNow) > 2.days {
                 lastRefreshTime.text = L10n.profileLastAppRefresh(TimeFormatter.shared.appleStyleElapsedString(date: lastUpdateTime))
                 alertIcon.isHidden = false
@@ -249,7 +295,7 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
                 alertIcon.isHidden = true
             }
         } else {
-            refreshBtn.buttonTitle = L10n.refreshNow
+            refreshButtonTitle = L10n.refreshNow
             lastRefreshTime.text = L10n.refreshPreviousRun(L10n.timeFormatNever)
             alertIcon.isHidden = false
         }
@@ -308,7 +354,6 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
 
         cell.settingsImage.tintColor = ThemeColor.primaryIcon01()
         cell.settingsLabel.setLetterSpacing(-0.01)
-        cell.updateImageScale()
         cell.separatorInset = .zero
 
         switch row {
@@ -468,8 +513,22 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
     }
 
     private func updateFooterFrame() {
-        let height: CGFloat = plusInfoView.isHidden ? 120 : 308
+        footerView.setNeedsLayout()
+        footerView.layoutIfNeeded()
+
+        let targetSize = CGSize(width: profileTable.bounds.width, height: UIView.layoutFittingCompressedSize.height)
+        let height = footerView.systemLayoutSizeFitting(targetSize, withHorizontalFittingPriority: .required, verticalFittingPriority: .fittingSizeLevel).height
+
         footerView.frame = CGRect(x: footerView.frame.minX, y: footerView.frame.minY, width: footerView.frame.width, height: height)
+        profileTable.tableFooterView = footerView
+    }
+
+    override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
+        super.traitCollectionDidChange(previousTraitCollection)
+
+        if traitCollection.preferredContentSizeCategory != previousTraitCollection?.preferredContentSizeCategory {
+            updateFooterFrame()
+        }
     }
 
     // MARK: - What's New Autoplay flow